refactor: Open Graph add
